helpers: move mailauth lib to separate dir
